feat: add ToolExecutionModeControl component for managing execution modes
